requirements.txt
beautifulsoup4~=4.12.3
selenium~=4.25.0
PyYAML~=6.0.2
